smash/requirements.txt
pychallonge==1.11.5
PyQt5==5.15.7
